feat: redesign billing page with 3-zone layout and tab navigation
Reorganize billing page UX: hero AccountStatusCard on top showing current
plan + quota, segmented tab bar for plans/replays/addons, and payment
management footer. Sync active tab to URL params for deep linking.
